fix: explicit types

+++ crates/actors/src/packing.rs
@@ -194,7 +194,8 @@ impl PackingActor {
                             chunk_range_split.start(),
                             chunk_range_split.end()
                         );
-                        let num_chunks = chunk_range_split.end() - chunk_range_split.start() + 1;
+                        let num_chunks =
+                            (chunk_range_split.end() - chunk_range_split.start() + 1) as u32;
                         let storage_module = storage_module.clone();
                         let chunk_range_split = chunk_range_split.clone();
